+++ src/client/java/dev/charles/multigolem/client/render/GolemRenderStateExtension.java
@@ -1,0 +1,8 @@
+package dev.charles.multigolem.client.render;
+
+import dev.charles.multigolem.GolemVariant;
+
+public interface GolemRenderStateExtension {
+    GolemVariant multigolem$getVariant();
+    void multigolem$setVariant(GolemVariant variant);
+}